Split out wasm PageInfo, add it to parse().
