+++ ftml/src/wasm/parsing.rs
@@ -18,6 +18,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
+use super::page_info::PageInfo;
 use super::prelude::*;
 use super::tokenizer::Tokenization;
 use crate::parsing::{
@@ -125,12 +126,13 @@ impl SyntaxTree {
 // Exported functions
 
 #[wasm_bindgen]
-pub fn parse(tokens: Tokenization) -> Result<ParseOutcome, JsValue> {
+pub fn parse(page_info: PageInfo, tokens: Tokenization) -> Result<ParseOutcome, JsValue> {
     let log = &*LOGGER;
 
     // Borrow and perform parsing
+    let page_info = page_info.get();
     let tokenization = tokens.get();
-    let (syntax_tree, warnings) = crate::parse(log, tokenization).into();
+    let (syntax_tree, warnings) = crate::parse(log, page_info, tokenization).into();
 
     // Deep-clone AST to make it owned, so it can be safely passed to JS.
     let syntax_tree = syntax_tree.to_owned();
